+++ assets/js/app.js
@@ -7,7 +7,7 @@ require('datatables.net-bs4')(window, $);
 require('datatables.net-fixedheader')();
 require('datatables.net-responsive-bs4')();
 // require('datatables.net-scroller-bs4')();
-// require('datatables.net-select-bs4')();
+require('datatables.net-select')();
 
 // Import local files
 import { configPublicChannel } from './socket';